# perf(container): decreased number of mutexes