`make benchmark`

**before**
```
BenchmarkContainer_scopeDefault-10                               8241607               144.5 ns/op
BenchmarkContainer_scopeShared-10                               11158279               107.6 ns/op
BenchmarkContainer_scopeContextual-10                            1565611               780.4 ns/op
BenchmarkContainer_scopeContextual_in_same_context-10           19036338                64.77 ns/op
BenchmarkContainer_scopeNonShared-10                             1730998               674.9 ns/op
BenchmarkContainer_map-10                                       470766387                2.455 ns/op
```

**after**
```
BenchmarkContainer_scopeDefault-10                               9019424               113.5 ns/op
BenchmarkContainer_scopeShared-10                               10615401               116.8 ns/op
BenchmarkContainer_scopeContextual-10                            1625832               798.5 ns/op
BenchmarkContainer_scopeContextual_in_same_context-10           17719750                66.04 ns/op
BenchmarkContainer_scopeNonShared-10                             1925256               619.4 ns/op
BenchmarkContainer_map-10                                       491388753                2.481 ns/op
```